+++ src/Exception/FilterQuery/FilterQueryException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Exception\FilterQuery;
+
+use App\Exception\KSearchException;
+
+abstract class FilterQueryException extends KSearchException
+{
+}